여정 삭제 Spring Security 기능 추가

+++ src/main/java/kr/co/fastcampus/travel/domain/itinerary/service/ItineraryService.java
@@ -27,8 +27,12 @@ public class ItineraryService {
         return ItineraryDto.from(itinerary);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id, String email) {
         var itinerary = findById(id);
+        String writerEmail = itinerary.getTrip().getMember().getEmail();
+        if (!writerEmail.equals(email)) {
+            throw new MemberMismatchException();
+        }
         itineraryRepository.delete(itinerary);
     }
